Changed location for launching program to pyMARS.py and made it so hd5f files and csv files are always saved.
